# Optimize `regexp_replace` when the input is a sparse array